feat: Refactor getUser to getUserById and deleteUser to deleteUserById

+++ server/controllers/user.controller.js
@@ -50,12 +50,12 @@ const getUsers = async (req, res, next) => {
   }
 };
 
-const getUser = async (req, res, next) => {
+const getUserById = async (req, res, next) => {
   try {
     const id = req.params.id;
     const options = { password: 0 };
-    const user = await findWithId(id, options);
-      
+    const user = await findWithId(User, id, options);
+
     if (!user) {
       return errorHandler(res, { statusCode: 404, message: "User not found" });
     } else {
@@ -63,7 +63,7 @@ const getUser = async (req, res, next) => {
         statusCode: 200,
         message: "User found successfully",
         payload: {
-          user
+          user,
         },
       });
     }
@@ -72,36 +72,33 @@ const getUser = async (req, res, next) => {
   }
 };
 
-const deleteUser = async (req, res, next) => {
+const deleteUserById = async (req, res, next) => {
   try {
     const id = req.params.id;
     const options = { password: 0 };
-    const user = await findWithId(id, options);
+    const user = await findWithId(User, id, options);
 
     const userImagePath = user.image;
-    fs.access(userImagePath, (err)=>{
-      if(err){
-        console.error("Image not found")
-      }else{
-        fs.unlink(userImagePath,(err)=>{
-          if(err) throw err;
-          console.log("Image Deleted Successfully")
-        })
+    fs.access(userImagePath, (err) => {
+      if (err) {
+        console.error("Image not found");
+      } else {
+        fs.unlink(userImagePath, (err) => {
+          if (err) throw err;
+          console.log("Image Deleted Successfully");
+        });
       }
     });
 
-    await User.findByIdAndDelete({ _id: id ,
-      isAdmin:false
-    });
+    await User.findByIdAndDelete({ _id: id, isAdmin: false });
 
-      return successHandler(res, {
-        statusCode: 200,
-        message: "User deleted successfully",
-      });
+    return successHandler(res, {
+      statusCode: 200,
+      message: "User deleted successfully",
+    });
   } catch (error) {
     next(error);
   }
 };
 
-
-export { getUsers, getUser, deleteUser };
+export { getUsers, getUserById, deleteUserById };
